Added constructors to sprite subclasses and made Animation a frontend class that contains and sets imageviews

+++ src/game_engine/sprite/Enemy.java
@@ -1,6 +1,6 @@
 package game_engine.sprite;
 
-import game_engine.Animation;
+import game_engine.game_player.Animation;
 
 public class Enemy extends Sprite {
 
@@ -14,7 +14,7 @@ public class Enemy extends Sprite {
         super(name);
     }
   
-    public Enemy(String name, double id){
+    public Enemy(String name, int id){
         super(name,id);
     }
 	@Override
